Reorganize [suggest] code
